moved SteppyError to separate py file (#37)
* moved SteppyError to separate py file

* prepare for 0.1.11
